Save shopify carts
This code sets up a Flask app with a webhook endpoint at /sms, which Twilio will use to send incoming text messages. When a message is received, the code checks if it contains a valid Shopify cart URL. If it does, it sends a request to the Shopify API to retrieve the cart data, and sends a confirmation message to the customer. If the cart data cannot be retrieved, it sends an error message. If the message does not contain a valid cart URL, it sends a reminder message to the customer to send a valid URL.